+++ src/plugin-marklogic/main/uk/co/reecedunn/intellij/plugin/marklogic/rewriter/lang/ModelTypeRegexLanguageInjection.kt
@@ -20,17 +20,20 @@ import com.intellij.lang.injection.MultiHostRegistrar
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiLanguageInjectionHost
 import com.intellij.psi.xml.XmlAttributeValue
-import com.intellij.psi.xml.XmlTag
 import org.intellij.lang.regexp.RegExpLanguage
-import uk.co.reecedunn.intellij.plugin.core.psi.contextOfType
-import uk.co.reecedunn.intellij.plugin.core.xml.attribute
+import uk.co.reecedunn.intellij.plugin.xdm.xml.XmlAccessorsProvider
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathStringLiteral
+import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryDirAttributeValue
 
 class ModelTypeRegexLanguageInjection : MultiHostInjector {
     override fun elementsToInjectIn(): MutableList<out Class<out PsiElement>> = ELEMENTS_TO_INJECT_IN
 
     override fun getLanguagesToInject(registrar: MultiHostRegistrar, context: PsiElement) {
-        val attribute = context.contextOfType<XmlAttributeValue>(false)?.attribute ?: return
-        if (attribute.localName != "matches" || !isModelTypeRegex(attribute.parent)) return
+        val (attribute, accessors) = XmlAccessorsProvider.attribute(context) ?: return
+        if (!accessors.hasNodeName(attribute, "", "matches")) return
+
+        val element = accessors.parent(attribute) ?: return
+        if (!accessors.hasNodeName(element, Rewriter.NAMESPACE, MODEL_TYPE_LOCAL_NAMES)) return
 
         val host = context as PsiLanguageInjectionHost
         val range = host.textRange
@@ -40,13 +43,13 @@ class ModelTypeRegexLanguageInjection : MultiHostInjector {
         registrar.doneInjecting()
     }
 
-    private fun isModelTypeRegex(tag: XmlTag): Boolean {
-        return tag.namespace == Rewriter.NAMESPACE && MODEL_TYPE_LOCAL_NAMES.contains(tag.localName)
-    }
-
     companion object {
         private val MODEL_TYPE_LOCAL_NAMES = setOf("match-header", "match-path", "match-string")
 
-        private val ELEMENTS_TO_INJECT_IN = mutableListOf(XmlAttributeValue::class.java)
+        private val ELEMENTS_TO_INJECT_IN = mutableListOf(
+            XmlAttributeValue::class.java,
+            XPathStringLiteral::class.java,
+            XQueryDirAttributeValue::class.java
+        )
     }
 }